Moved Models to a seperate file, models.py.
