+++ symfit/api.py
@@ -2,9 +2,10 @@
 import symfit.core.operators
 
 # Expose useful objects.
-from symfit.core.fit import (
-    Fit, Model, ODEModel, ModelError, CallableModel,
-    CallableNumericalModel, GradientModel
+from symfit.core.fit import Fit
+from symfit.core.models import (
+    Model, ODEModel, ModelError, CallableModel, CallableNumericalModel,
+    GradientModel
 )
 from symfit.core.fit_results import FitResults
 from symfit.core.argument import Variable, Parameter
